Refactor main library
